Finished 04_removeFromArray
This task was to work with variadic functions where
the function has two input parameters where one
is an array and the second one is a rest parameter.
The goal of the task to output an array that only
had values which were not common between the two
arrays. The task was done successfully and the
desired output was met. All tests were successfull.