Finished 05_sumAll exercise
The task was to create a function which accepted
two positive integer parameters in order to generate
the integer values inbetween the two numbers and then
sum all of them. The task utilized the concepts of
for loops to generate the numbers and to add them all.
All the test cases were successful and the proper
output was acheived.